+++ src/components/todo/TodoItem.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useRef, useEffect, useContext } from 'react';
 import useAxios from '../../hooks/useAxios';
-import { todoAPI } from '../../store/api';
+import { todoAPI } from '../../utils/api';
 import { TodosDispatchContext } from '../../store/todo-context';
 import Input from '../UI/Input';
 import classes from './TodoItem.module.css';
